refactor: move GetOrCreateBundleFileIdAsync to vrchat api client
